backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.1
Flask-Limiter==3.7.0
psycopg2-binary==2.9.9
redis==5.0.7
celery==5.4.0
marshmallow==3.21.3
argon2-cffi==23.1.0
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
gunicorn==22.0.0
